requirements.txt
aiogram
python-dotenv
asyncio
sqlalchemy
aiosqlite
asyncpg